requirements.txt
requests
mss
Pillow
numpy
opencv-python
edge-tts
pygame
SpeechRecognition
sounddevice
scipy
pyserial
paho-mqtt
colorama
python-dotenv
PyAudio
